Various edits to the final tool in the toolbox
Code fails if path to geodatabase has a hyphen in it, the code now traps for this situations and fails gracefully. Code saves the adjusted raster and adds it to the map calling it "adjusted". Improved messaging and cleared out unused commented out lines of code to make it easier to understand what's happening